+++ backend_rust/src/adapters.rs
@@ -1,4 +1,4 @@
-pub mod db;
 pub mod grpc;
 pub mod recipe_mapper;
+pub mod repository;
 pub mod step_mapper;
