+++ backend_rust/src/adapters/repository.rs
