+++ backend_rust/src/adapters/repository/mysql/entity/process_registrations.rs
