+++ backend_rust/src/adapters/repository/mysql/entity/recipes.rs
