+++ backend_rust/src/application.rs
@@ -1,4 +1,3 @@
 pub mod mapper;
 pub mod port;
-pub mod repository;
 pub mod usecase;
