+++ backend_rust/src/application/port.rs
@@ -1,3 +1,4 @@
 pub mod external;
+pub mod repository;
 
 pub use self::external::process_service_port::ProcessServicePort;
