+++ backend_rust/src/application/port/repository/resource_repository.rs
